Added comments to explain the booting for GDT
