+++ src/kernel/arch/x86/gdt.rs
@@ -1,10 +1,29 @@
-use super::{
-	boot::{GdtGates, GDT_ENTRIES},
-	diagnostics::cpu::check_protection_status,
-};
+//------------------------------------------------------------------------------
+// Global Descriptor Table Implementation (gdt.rs)
+//
+// This module implements the GDT entry structure and manipulation methods.
+// Each GDT entry (Gate) is a 64-bit structure with the following layout:
+//
+// Base Address:  32 bits (split across bits 16-39 and 56-63)
+// Segment Limit: 20 bits (bits 0-15 and 48-51)
+// Access Byte:   8 bits  (bits 40-47)
+//   - Present:     Bit 7 (P)
+//   - DPL:         Bits 5-6 (Ring Level)
+//   - Type:        Bit 4 (S)
+//   - Type flags:  Bits 0-3
+// Flags:        4 bits  (bits 52-55)
+//   - Granularity: Bit 3 (G)
+//   - Size:        Bit 2 (D/B)
+//   - Long mode:   Bit 1 (L)
+//   - Reserved:    Bit 0
+//
+// For more information go to:
+// https://wiki.osdev.org/Global_Descriptor_Table
+//------------------------------------------------------------------------------
 
 pub struct Gate(pub u64);
 
+/// Must be packed to maintain exact CPU-required layout
 #[repr(C, packed)]
 pub struct GDTDescriptor {
 	pub size: u16,
@@ -13,6 +32,13 @@ pub struct GDTDescriptor {
 
 #[allow(unused)]
 impl Gate {
+	/// Creates a new GDT entry with specified parameters
+	///
+	/// # Arguments
+	/// * `base` - 32-bit base address of the segment
+	/// * `limit` - 20-bit size of the segment
+	/// * `access` - 8-bit access flags (present, DPL, type)
+	/// * `flags` - 4-bit flags (granularity, size, long mode)
 	pub const fn new(base: u32, limit: u32, access: u8, flags: u8) -> Self {
 		let mut c = Self(0);
 		c.set_base(base);
@@ -64,22 +90,4 @@ impl Gate {
 		self.0 &= !(0xf << 52);
 		self.0 |= (flags as u64) << 52;
 	}
-}
-
-extern "C" {
-	fn gdt_flush(gdt_ptr: *const GDTDescriptor);
-}
-
-#[no_mangle]
-pub fn gdt_init() {
-	let gdt_descriptor = GDTDescriptor {
-		size: (core::mem::size_of::<GdtGates>() - 1) as u16,
-		offset: &GDT_ENTRIES as *const _ as u32,
-	};
-
-	unsafe {
-		gdt_flush(&gdt_descriptor as *const _);
-	}
-
-	check_protection_status();
 }
